// src/app/services/projects/models/update-task-request.ts
/* tslint:disable */
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

export interface UpdateTaskRequest {
  assigneeId?: string;
  description?: string;
  dueDate?: string;
  priority?: 'LOW' | 'MEDIUM' | 'HIGH' | 'CRITICAL';
  sprintId?: string;
  status?: 'TODO' | 'IN_PROGRESS' | 'IN_REVIEW' | 'DONE' | 'BLOCKED';
  storyPoints?: number;
  taskType?: 'TASK' | 'BUG' | 'STORY' | 'EPIC' | 'SUBTASK';
  title?: string;
}